+++ CHANGES.rst
@@ -4,6 +4,10 @@ Changelog
 1.3.0 (unreleased)
 ------------------
 
+New features:
+
+- `Sublime​Linter linter arguments are possible <http://www.sublimelinter.com/en/latest/linter_settings.html#args>`_ to provide. Currently ``sublimelinter-flake8`` and ``sublimelinter-pylint`` are supported.
+
 - `New style Sublime​Linter linter settings <http://www.sublimelinter.com/en/stable/settings.html#project-settings>`_ generation has been added. [nazrulworld]